Lynx support: Open browser asynchronous with current tty as stdin/stdout
This patch enables the option to login using the minimalistic console
browser Lynx, if xdg-open is running it as default browser.  Useful on
remote machines without X-forwarding or VNC.

How it works
************

Browsers stdin/stdout are outputting/reading to/from TTY if OS is
Linux.

Browsers are executed asynchronous, then git-credential-oauth is
waiting for the response of the temporary HTTP server, followed by
WAIT() until the browser was closed (or forked if GUI is available).